Tweak monitoring code to support collectd
This is really a local patch for my own lab environment. The start/stop
functions launch and then kill collectd on each node. The nodes send their
data to a common collection machine (running graphite), and right now
there is no provision to copy the collected data back to the CBT archive
directory.